some key set details work in progress
